array indexing by a range

+++ src/range.c
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 #include "memory.h"
 
-range_t *range_new(double start, double end)
+range_t *range_new(long start, long end)
 {
   range_t *range = (range_t *) allocate(sizeof(*range));
   range->ref_count = 0;
@@ -24,7 +24,7 @@ void range_free(range_t *range)
 
 void range_print(range_t *range)
 {
-  printf("%g..%g", range->start, range->end);
+  printf("%ld..%ld", range->start, range->end);
 }
 
 bool range_equal(range_t *range1, range_t *range2)
